device_controller/requirements.txt
requests
astral
noaa_sdk
schedule
time-machine
pytz
firebase_admin
firebase-rest-api
pyjwt